feat: support specifying type of props for styled with TypeScript
Currently we infer the props for the components based on the type of the base component, e.g. - for `styled.button`, it'll be the props of `button`, for `styled(Title)`, it'll be the props of `Title` etc. However, it's common to pass additional props to a styled component specifically for styling. For example:

```js
const Button = styled.button`
  background: $(props => props.primary ? 'blue' : 'white'};
`;
```

Here the user needs to pass a `primary` prop even if `button` doesn't accept it. Currently we workaround this by allowing the user to pass any extra props. However the extra props will be typed as `any`, which is not exactly type cheking.

TypeScript has this cool feature of being able to have optional type parameters, which we can use here to allow the user to specify the types manually, while using the inferred types by default when a prop is not specified. So it'll basically look like this:

```js
const Button = styled.button<{ primary: boolean; children: string }>`
  background: $(props => props.primary ? 'blue' : 'white'};
`;
```