fix(esbuild): workaround to for esbuild compile options (#784)
When using esbuild with preact the jsxFactory and jsxFragment
parameter are used to use h() rather than React.createElement(). This
change ensures that the parameters are passed to the transform
function.

NOTE 1: Usually esbuild would just pick up on tsconfig.json - this
workaround won't work there. The pararmeters must be provided
explicitly.

NOTE 2: The plugin should probably be rewritten on a more fundamental
level as it interefers heavily with the esbuild build process.